Update Componen
Menambah components dan memperbaiki code
